fix(config): load local .env and align local dev defaults
- auto-load .env in config.Load so go run uses the same local overrides as documentation expects
- restore config fields used by router/runtime with safe defaults and legacy metrics fallback to keep builds stable
- clarify README local development paths and env templates so host-run Go and docker-compose mail settings are not mixed

Release note: Fixed local go-run configuration loading and refreshed local development docs for clearer setup paths.

+++ internal/router/router.go
@@ -164,7 +164,7 @@ func New(
 		ticker := time.NewTicker(1 * time.Hour)
 		defer ticker.Stop()
 		for range ticker.C {
-			deleted, err := deleteSvc.CleanExpiredImages(context.Background(), cfg.Server.ExpiredCleanupBatchSize)
+			deleted, err := deleteSvc.CleanExpiredImages(context.Background(), int32(cfg.Server.ExpiredCleanupBatchSize))
 			if err != nil {
 				slog.Warn("failed to clean expired images", "error", err)
 			} else if deleted > 0 {
